backend/requirements.txt
# Finance Agent API - Python Dependencies

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Google Gemini AI
google-generativeai==0.3.2
google-genai==0.3.0

# Data Models & Validation
pydantic==2.6.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Client (for future external API calls)
httpx==0.26.0

# Date/Time Utilities
python-dateutil==2.8.2
